requirements.txt
absl-py==0.8.0
astor==0.8.0
certifi==2019.9.11
cycler==0.10.0
gast==0.3.2
grpcio==1.24.1
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.14.5
opencv-python==3.4.0.14
Pillow==6.2.0
protobuf==3.10.0
pyparsing==2.4.2
python-dateutil==2.8.0
six==1.12.0
tensorboard==1.10.0
tensorflow-gpu==1.10.0
termcolor==1.1.0
Werkzeug==0.16.0
wincertstore==0.2
